% Generated by roxygen2 (4.0.1): do not edit by hand
\name{integrand}
\alias{integrand}
\title{Evaluate integrand alone}
\usage{
integrand(z, mat1)
}
\arguments{
\item{z}{This is a real number between -Inf and Inf.}

\item{mat1}{This is the matrix of the delta values. This matrix can be
generated using the \code{\link{genDelMat}} function.}
}
\value{
The function returns a scalar value.
}
\description{
This function will evaluate the integrand in the expression for the exact
coverage probability.
}
\seealso{
\code{\link{exactCoverageProb}}, \code{\link{integrate2}}
}

