% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optimalLambda}
\alias{optimalLambda}
\title{Derive the optimal lambda}
\usage{
optimalLambda(alpha, n, p, k = 1, var.known = TRUE)
}
\arguments{
\item{alpha}{The desired confidence coefficient.}

\item{n}{The number of replications per population.}

\item{p}{The number of populations considered. This must be present if
min.loc is equal to 'zero'.}

\item{k}{The number of populations selected.}

\item{var.known}{A logical flag indicating if the variance of the
observations is known exactly. It is TRUE by default.}
}
\value{
The function returns a scalar value.
}
\description{
Derives the optimal lambda for a given alpha.
}
\details{
This will find the optimal lambda to be used for the shrinkage of
confidence intervals.
}
\seealso{
\code{\link{optimalC}}
}

