% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optimalLambdaC}
\alias{optimalLambdaC}
\title{Derive the optimal lambda and c value}
\usage{
optimalLambdaC(alpha = 0.05, n, p, k = 1, var.known = TRUE, eps = 0.1)
}
\arguments{
\item{alpha}{The desired confidence coefficient.}

\item{n}{The number of replications per population.}

\item{p}{The number of populations considered. This must be present if
min.loc is equal to 'zero'.}

\item{k}{The number of populations selected.}

\item{var.known}{A logical flag indicating if the variance of the
observations is known exactly. It is TRUE by default.}

\item{eps}{The grid size that is to be set up.}
}
\value{
The function returns a list with two components, lambda and c.val
that are optimal.
}
\description{
Derives the optimal lambda and c-value for a given configuration.
}
\details{
This function will return the optimal lambda and c-value to be used,
using a grid search.

There are essentially 2 different cases to consider. They correspond to the
cases when the variance is known or unknown.
}

