% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{popKorn-package}
\alias{popKorn-package}
\title{Interval estimation for selected populations}
\description{
Interval estimation for selected populations
}
\details{
\tabular{ll}{
Package: \tab popKorn\cr
Type: \tab Package\cr
Version: \tab 0.3-0\cr
Date: \tab 2014-07-04\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}

This package provides routines for estimating the interval for the mean of
selected populations.
}
\author{
Claudio Fuentes, Vik Gopal
}
\keyword{package}

